+++ xnicer/xdeconv/__init__.py
@@ -1,3 +1,4 @@
+# pylint: disable=no-name-in-module
 from .em_step import FIX_NONE, FIX_AMP, FIX_CLASS, FIX_MEAN, FIX_COVAR, \
     FIX_ALL, em_step, log_likelihoods
 from .xdeconv import xdeconv