+++ xnicer/xnicer.py
@@ -13,6 +13,7 @@ import warnings
 import copy
 from scipy.special import ndtri, logsumexp
 from sklearn.base import BaseEstimator
+from astropy import table
 from .xdeconv import XD_Mixture, FIX_MEAN, FIX_COVAR
 from .utilities import log1mexp, cho_solve, cho_matrix_solve
 from .catalogs import ExtinctionCatalogue
@@ -28,12 +29,13 @@ class XNicer(BaseEstimator):
     Parameters
     ----------
     xdmix : XD_Mixture
-        An XD_Mixture instance used to perform all necessary extreme deconvolutions.
+        An XD_Mixture instance used to perform all necessary extreme 
+        deconvolutions.
 
     extinctions : array-like, shape (n_extinctions,)
         A 1D vector of extinctions used to perform a selection correction.
 
-    log_weights_ : tuple of array-like, shape (n_extinctions, xdmix.sum_components))
+    log_weights_ : tuple of array-like, shape (n_extinctions, xdmix.n_components))
         The log of the weights of the extreme decomposition, for each class of
         objects, at each extintion value.
     
@@ -68,9 +70,11 @@ class XNicer(BaseEstimator):
             If true, the first value of self.extinctions is skipped, which
             makes the whole procedure much faster.
         """
+        n_exts = len(self.extinctions)
+        use_classes = cat['log_class_probs'] is not None
         for n, extinction in enumerate(self.extinctions):
-            # If update is set, skip the first iteration: this makes the fitting
-            # procedure much faster
+            # If update is set, skip the first iteration: this makes the 
+            # fitting procedure much faster
             if update and n == 0:
                 continue
             # Add the extinction, as requested
@@ -93,14 +97,18 @@ class XNicer(BaseEstimator):
                 # We could set this earlier in the __init__, but it does not
                 # work in case the numbero of components for self.xdmix is an
                 # array (this is possible if we request a BIC criterion)
-                self.log_weights_ = np.zeros(
-                    (len(self.extinctions), self.xdmix.sum_components))
+                self.log_weights_ = np.zeros((n_exts, self.xdmix.n_components))
+                if use_classes:
+                    self.log_classes_ = np.zeros(
+                        (n_exts, self.xdmix.n_components, self.xdmix.n_classes))
             with np.errstate(divide='ignore'):
                 self.log_weights_[n] = np.log(self.xdmix.weights_)
+                if use_classes:
+                    self.log_classes_[n] = np.log(self.xdmix.classes_)
 
 
     def calibrate(self, cat, extinctions=None, **kw):
-        """Perform a full calibration of the algorithm for a set of test extinctions.
+        """Perform a full calibration of the algorithm for a set of extinctions.
         
         The calibration works by simulating in turn all extionctions provided,
         and by checking the final bias and (inverse) variances of the
@@ -146,7 +154,8 @@ class XNicer(BaseEstimator):
         self.calibration = (np.array(extinctions), np.array(biases),
                             np.array(ivars) / ivars[0])
             
-    def predict(self, cat, use_projection=True, full=False, n_iters=3):
+    def predict(self, cat, use_projection=True, use_classes=True, 
+                full=False, n_iters=3):
         """Compute the extinction for each object of a PhotometryCatalogue
 
         Parameters
@@ -156,6 +165,12 @@ class XNicer(BaseEstimator):
 
         use_projection : bool, default=True
             Parameter passed directly to `PhotometricCatalogue`.
+
+        use_classes : bool, default=True
+            If classes are available in the catalogue, and the deconvolution
+            has been performed using classes, takes them into account for
+            the extinction estimate. In this case, the final catalogue will
+            also have a column called 'log_class_probs'.
 
         full : bool, default=False
             If True, the function returns complete extinction catalogue,
@@ -170,16 +185,9 @@ class XNicer(BaseEstimator):
         cols = cat.get_colors(use_projection=use_projection)
         
         # Check if we need to use classes
-        if 'log_class_probs' in cols.colnames and \
-            isinstance(self.xdmix.n_components, tuple):
+        if use_classes and 'log_class_probs' in cols.colnames and \
+            self.xdmix.n_classes > 0:
             use_classes = True
-            # Distribute equally the class probabilities among the class members
-            full_class_prob = np.empty((len(cols), self.xdmix.sum_components))
-            cum_c = 0
-            for e, c in enumerate(self.xdmix.n_components):
-                full_class_prob[:, cum_c:cum_c +
-                                c] = (cols['log_class_probs'][:, e] - np.log(c))[:, np.newaxis]
-                cum_c += c
         else:
             use_classes = False
 
@@ -189,9 +197,13 @@ class XNicer(BaseEstimator):
         # this case, we need to keep track of the original objects and of the
         # associated probabilities.
         res = ExtinctionCatalogue()
-        res.meta['n_components'] = self.xdmix.sum_components,
-        res['idx'] = cols['idx']
+        res.meta['n_components'] = self.xdmix.n_components,
+        res.add_column(table.Column(
+            cols['idx'], name='idx',
+            description='index with respect to the original catalogue',
+            format='%d'))
         res.meta['n_colors'] = (cat.meta['n_bands']-1) if full else 0
+        res.meta['n_classes'] = (self.xdmix.n_classes) if use_classes else 0
 
         # Compute the extinction vector
         color_ext_vec = cols.meta['reddening_law']
@@ -217,17 +229,24 @@ class XNicer(BaseEstimator):
         T_d = cho_solve(Tc, d)
         Tlogdet = np.sum(np.log(np.diagonal(Tc, axis1=-2, axis2=-1)), axis=-1)
         sigma_k2 = 1.0 / np.sum(T_k * T_k, axis=-1)
-        # We are now ready to compute the means, variances, and weights for the
-        # GMM of the extinction. Each of these parameters has a shape
-        # (n_objs, n_bands - 1). We can go on with the means and variances, but
-        # the weights require more care, because they are linked to the xdmix
-        # weights, and these in turn depend on the particular extinction of the
-        # star. We therefore initially only compute log_weights0_, the log of
-        # the weights of the extinction GMM that does not include any term related
-        # to the xdmix weight.
-        res['means_A'] = sigma_k2 * np.sum(T_k * T_d, axis=-1)
-        # variances_ = np.sum(T_d * T_d, axis=-1) - self.ext*self.ext / sigma_k2
-        res['variances_A'] = sigma_k2
+        # We are now ready to compute the means, variances, and weights for
+        # the GMM of the extinction. Each of these parameters has a shape
+        # (n_objs, n_bands - 1). We can go on with the means and variances,
+        # but the weights require more care, because they are linked to the
+        # xdmix weights, and these in turn depend on the particular extinction
+        # of the star. We therefore initially only compute log_weights0_, the
+        # log of the weights of the extinction GMM that does not include any
+        # term related to the xdmix weight.
+        res.add_column(table.Column(
+            sigma_k2 * np.sum(T_k * T_d, axis=-1), name='means_A', 
+            description='Array of means of extinction components',
+            unit='mag', format='%6.3f'))
+        # variances_ = np.sum(T_d * T_d, axis=-1) - self.ext*self.ext /
+        # sigma_k2
+        res.add_column(table.Column(
+            sigma_k2, name='variances_A',
+            description='Array of variances of extinction components',
+            unit='mag^2', format='%7.5f'))
         C_k = np.sum(T_d * T_d, axis=-1) - res['means_A']**2 / sigma_k2
         log_weights0_ = - Tlogdet - \
             (cat.meta['n_bands'] - 1) * np.log(2.0*np.pi) / 2.0 - \
@@ -236,30 +255,56 @@ class XNicer(BaseEstimator):
         if full:
             T_V = cho_matrix_solve(Tc, PV)
             V_TT_k = np.einsum('...ji,...j->...i', T_V, T_k)
-            res['covariances'] = -V_TT_k * res['variances_A'][:, :, np.newaxis]
-            res['variances_c'] = V - np.einsum('...ij,...ik->...jk', T_V, T_V) - \
-                np.einsum('...i,...j->...ij', V_TT_k, res['covariances'])
-            res['means_c'] = self.xdmix.means_[np.newaxis, :, :] + \
+            res.add_column(table.Column(
+                -V_TT_k * res['variances_A'][:, :, np.newaxis], 
+                name='covariances',
+                description='Array of covariances of extinction-magnitude',
+                unit='mag^2', format='%7.5f'))
+            res.add_column(table.Column(
+                V - np.einsum('...ij,...ik->...jk', T_V, T_V) -
+                np.einsum('...i,...j->...ij', V_TT_k, res['covariances']),
+                name='variances_c',
+                description='Array of variances of intrinsic magnitudes',
+                unit='mag^2', format='%7.5f'))
+            res.add_column(table.Column(
+                self.xdmix.means_[np.newaxis, :, :] +
                 np.einsum('...ji,...j->...i', T_V, T_d) - \
-                V_TT_k * res['means_A'][:,:, np.newaxis]
+                V_TT_k * res['means_A'][:,:, np.newaxis],
+                name='means_c',
+                description='Array of means of intrinsic magnitudes',
+                unit='mag', format='%6.3f'))
         # Fine, now need to perform two loops: one outer loop where we iterate
-        # n_iters time, and an internal loop that computes the weights for each
-        # extinction step. To proceed we define a new array, log_ext_weights,
-        # that saves the log of the contribution of a particular extinction step
-        # for each object; we initially assume that all objects only have a
-        # contribution from the first extinction step.
+        # n_iters time, and an internal loop that computes the weights for
+        # each extinction step. To proceed we define a new array,
+        # log_ext_weights, that saves the log of the contribution of a
+        # particular extinction step for each object; we initially assume that
+        # all objects only have a contribution from the first extinction step.
         log_ext_weights = np.full(
             (len(cols), len(self.extinctions)), -np.inf)
         log_ext_weights[:, 0] = 0.0
         for _ in range(n_iters):
-            res['log_weights'] = log_weights0_ + \
-                logsumexp(self.log_weights_[
-                            np.newaxis, :, :] + log_ext_weights[:, :, np.newaxis], axis=1)
             # If the classes are available, use them
             if use_classes:
-                res['log_weights'] += full_class_prob
-            res['log_evidence'] = logsumexp(res['log_weights'], axis=-1)
-            res['log_weights'] -= res['log_evidence'][..., np.newaxis]
+                # The logsumexp's stuff has shape (n, e, k, c), where n=# objs,
+                # e=# calibration extinctions, k=# clusters, and c=# classes
+                # I am summing over all extinctions to use the correct
+                # combination of values.
+                tmp = log_weights0_[:, :, np.newaxis] + logsumexp(
+                    self.log_weights_[np.newaxis, :, :, np.newaxis] + 
+                    self.log_classes_[np.newaxis, :, :, :] + 
+                    cols['log_class_probs'][:, np.newaxis, np.newaxis, :] + 
+                    log_ext_weights[:, :, np.newaxis, np.newaxis], axis=1)
+                res['log_weights'] = logsumexp(tmp, axis=2)
+                res['log_class_probs'] = logsumexp(tmp, axis=1)
+                res['log_evidence'] = logsumexp(res['log_weights'], axis=-1)
+                res['log_weights'] -= res['log_evidence'][:, np.newaxis]
+                res['log_class_probs'] -= res['log_evidence'][:, np.newaxis]
+            else:
+                res['log_weights'] = log_weights0_ + logsumexp(
+                    self.log_weights_[np.newaxis, :, :] + 
+                    log_ext_weights[:, :, np.newaxis], axis=1)
+                res['log_evidence'] = logsumexp(res['log_weights'], axis=-1)
+                res['log_weights'] -= res['log_evidence'][:, np.newaxis]
             # Now we need to update the weights for the extinction steps according
             # to each object's average extinction
             for e, extinction in enumerate(self.extinctions):
@@ -267,6 +312,21 @@ class XNicer(BaseEstimator):
                     np.repeat(extinction, len(cols)), np.zeros(len(cols)))
             log_ext_weights -= logsumexp(log_ext_weights, axis=-1)[..., np.newaxis]
         res.update_()
+        res['mean_A'].description = 'Mean of the posterior extinction'
+        res['mean_A'].unit = 'mag'
+        res['mean_A'].format = '%6.3f'
+        res['variance_A'].description = 'Variance of the posterior extinction'
+        res['variance_A'].unit = 'mag^2'
+        res['variance_A'].format = '%7.3f'
+        res['log_weight'].description = 'Final log of the weight'
+        res['log_weights'].description = 'log of the weight of each component'
+        res['log_evidence'].description = 'log of the evidence'
+        res['log_weight'].format = res['log_weights'].format = \
+            res['log_evidence'].format = '%5g'
+        if use_classes:
+            res['log_class_probs'].description = \
+                'log of the probability to belong to each class'
+            res['log_class_probs'].format = '%g'
 
         # In case we have a calibrated object, perform the bias correction and
         # the XNicest estimates
@@ -291,11 +351,19 @@ class XNicer(BaseEstimator):
             log_ext_weights -= logsumexp(log_ext_weights, axis=-1)[..., np.newaxis]
             ext_weights = np.exp(log_ext_weights)
             # Finally perform the XNicest evaluations
-            res['xnicest_weight'] = np.sum(
-                ext_weights / self.calibration[2], axis=1)
-            res['xnicest_bias'] = np.sum(ext_weights * self.calibration[0]
-                / self.calibration[2], axis=1) / res['xnicest_weight'] - np.sum(
-                    ext_weights * self.calibration[0], axis=1)
+            res.add_column(table.Column(np.sum(
+                ext_weights / self.calibration[2], axis=1),
+                name='xnicest_weight',
+                description='Object\'s weight to use with the XNicest alogorithm',
+                format='%7.5f'))
+            res.add_column(table.Column(
+                np.sum(ext_weights * self.calibration[0]
+                       / self.calibration[2], axis=1) / 
+                res['xnicest_weight'] - np.sum(
+                    ext_weights * self.calibration[0], axis=1),
+                name='xnicest_bias',
+                description='Bias to subtract with the XNicest alogorithm',
+                format='%7.5f'))
         # Now, in case of use of log_probs, we need to correct the weights so
         # to include the original color weights.
         if 'log_probs' in cols.colnames: